Remove I/O jumpers
The I/O jumpers seemed to likely to either short a chip on baord damaging the driver, or cause confusion where someone forgets the state of a jumper and the inputs do not behave as expected.  Also a loose solder ball could also cause unexpected behavior.  As such, the risk is much greater than any reward (i.e. cost savings in fewer chips needed, and convenience in not having to externally pool inputs).