Removed LED combine jumpers
I removed the LED combining jumpers as the only case where they would be helpgul is if the driver is in a high current/power mode.  In this state, parallel wiring is likely better anyway to reduce thermal load on the PCR (compared to sodler only conduction in the case of a solder jumper), and their removal removed the risk of shorting between LEDs due to a solder ball.