Put current output inside voltage divider
If the cirrent sense voltage is directly sent externally, and accidental short would both cause full current to go through that channel (potentially destroying the LED and driver channel) but also through the external device potentially destroying it too.  Therefore, by putting it inside the divider, the current sense voltage is isolated from the outside.

The impedance of the divider was therefore also reduced to 1k, so that 10 MHz signals will be minimally attenuated.