Find driver using WIM
This code allows for the Teensy to be assigned a custom serial number that will show up on the WIM.  This way, the Python script can simply search all connected WIMs for ones that have the right product ID and serial number.   This means that there is no need for magic bytes that could corrupt other devices!