Seq upload bug fixed
Timeout bug was because small packets would reduce the timeout with the dynamic timeout feature used for packets.  This was fixed by setting a master default timeout, and dynamic timeouts can only be larger than this.  The Recv seq also resets the dynamic timeout to the default for header packets.